dovrseni editioni

+++ Backend/guide/src/main/java/com/example/guide/service/impl/BookServiceJpa.java
@@ -1,17 +1,31 @@
 package com.example.guide.service.impl;
 
+import com.example.guide.controller.responses.BookDeletionException;
 import com.example.guide.domain.Book;
+import com.example.guide.domain.BookEdition;
+import com.example.guide.domain.Edition;
+import com.example.guide.domain.EditionOrder;
 import com.example.guide.dto.BookDTO;
+import com.example.guide.repository.BookEditionRepository;
 import com.example.guide.repository.BookRepository;
+import com.example.guide.repository.EditionOrderRepository;
 import com.example.guide.service.BookService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
 public class BookServiceJpa implements BookService {
     @Autowired
     private BookRepository bookRepo;
+
+    @Autowired
+    private BookEditionRepository bookEditionRepo;
+
+    @Autowired
+    private EditionOrderRepository editionOrderRepo;
 
     @Override
     public List<Book> listAll() {
@@ -47,6 +61,10 @@ public class BookServiceJpa implements BookService {
     public void deleteBook(Long bookId) {
         Book book = bookRepo.findById(bookId)
                 .orElseThrow(() -> new RuntimeException("Book not found with ID: " + bookId));
+        List<BookEdition> bookEditions = bookEditionRepo.findAllByBook(book);
+        if (!bookEditions.isEmpty()) {
+            throw new BookDeletionException("Cant delete books with editions!");
+        }
         bookRepo.delete(book);
     }
 }
